Fix infinite fighting bug && play again!

+++ Assets/Scripts/Card.cs
@@ -171,7 +171,7 @@ public class Card {
 		}, null, null),
 		// Tier 2
 		new Card ("Will", "Draw extra cards equal to your level + 1 and play again", Color.GREEN, true, (Card augmentation, Card other, PlayerScript player, PlayerScript opponent) => {
-			player.Message(GameManagerScript.MESSAGE.DRAW, 1 + Mathf.Floor(player.school.advancement / 6));
+			player.Message(GameManagerScript.MESSAGE.DRAW, (int) (1 + Mathf.Floor(player.school.advancement / 6)));
 		}, null, null, null),
 		new Card ("Mind Swap", "Switch augmentations with your opponent", Color.BLUE, false, (Card augmentation, Card other, PlayerScript player, PlayerScript opponent) => {
 			Card temp = augmentation.previous;
